+++ examples/server_auth.py
@@ -4,6 +4,7 @@ Example "auth" server
 Ask mojang to authenticate the user
 """
 
+from twisted.internet import reactor
 from quarry.net.server import ServerFactory, ServerProtocol
 
 
@@ -12,7 +13,7 @@ class AuthProtocol(ServerProtocol):
         # This method gets called when a player successfully joins the server.
         #   If we're in online mode (the default), this means auth with the
         #   session server was successful and the user definitely owns the
-        #   username they claim to.
+        #   display name they claim to.
 
         # Call super. This switches us to "play" mode, marks the player as
         #   in-game, and does some logging.
@@ -20,9 +21,9 @@ class AuthProtocol(ServerProtocol):
 
         # Define your own logic here. It could be an HTTP request to an API,
         #   or perhaps an update to a database table.
-        username = self.username
+        display_name = self.display_name
         ip_addr = self.remote_addr.host
-        self.logger.info("[%s authed with IP %s]" % (username, ip_addr))
+        self.logger.info("[%s authed with IP %s]" % (display_name, ip_addr))
 
         # Kick the player.
         self.close("Thanks, you are now registered!")
@@ -30,28 +31,23 @@ class AuthProtocol(ServerProtocol):
 
 class AuthFactory(ServerFactory):
     protocol = AuthProtocol
+    motd = "Auth Server"
 
 
-def main(args):
+def main(argv):
     # Parse options
-    import optparse
-    parser = optparse.OptionParser(
-        usage="usage: %prog [options]")
-    parser.add_option("-a", "--host",
-                      dest="host", default="",
-                      help="address to listen on")
-    parser.add_option("-p", "--port",
-                      dest="port", default="25565", type="int",
-                      help="port to listen on")
-    (options, args) = parser.parse_args(args)
+    import argparse
+    parser = argparse.ArgumentParser()
+    parser.add_argument("-a", "--host", default="", help="address to listen on")
+    parser.add_argument("-p", "--port", default=25565, type=int, help="port to listen on")
+    args = parser.parse_args(argv)
 
     # Create factory
     factory = AuthFactory()
-    factory.motd = "Auth Server"
 
     # Listen
-    factory.listen(options.host, options.port)
-    factory.run()
+    factory.listen(args.host, args.port)
+    reactor.run()
 
 
 if __name__ == "__main__":