+++ examples/server_downtime.py
@@ -4,6 +4,7 @@ Example "downtime" server
 When a user tries to connect, the server will kick them with the MOTD
 """
 
+from twisted.internet import reactor
 from quarry.net.server import ServerFactory, ServerProtocol
 
 
@@ -17,29 +18,23 @@ class DowntimeFactory(ServerFactory):
     protocol = DowntimeProtocol
 
 
-def main(args):
+def main(argv):
     # Parse options
-    import optparse
-    parser = optparse.OptionParser(
-        usage="usage: %prog [options]")
-    parser.add_option("-a", "--host",
-                      dest="host", default="",
-                      help="address to listen on")
-    parser.add_option("-p", "--port",
-                      dest="port", default="25565", type="int",
-                      help="port to listen on")
-    parser.add_option("-m", "--message",
-                      dest="message", default="We're down for maintenance",
-                      help="message to kick users with")
-    (options, args) = parser.parse_args(args)
+    import argparse
+    parser = argparse.ArgumentParser()
+    parser.add_argument("-a", "--host", default="", help="address to listen on")
+    parser.add_argument("-p", "--port", default=25565, type=int, help="port to listen on")
+    parser.add_argument("-m", "--message", default="We're down for maintenance",
+                        help="message to kick users with")
+    args = parser.parse_args(argv)
 
     # Create factory
     factory = DowntimeFactory()
-    factory.motd = options.message
+    factory.motd = args.message
 
     # Listen
-    factory.listen(options.host, options.port)
-    factory.run()
+    factory.listen(args.host, args.port)
+    reactor.run()
 
 
 if __name__ == "__main__":